refactor(core): move parseExpression in a dedicated file
	+ remove the use of the deprecated PublicodesInternalError
